requirements.txt
fabric~=3.2.2
pyserial~=3.5
mobly~=1.12.2
PyYAML~=6.0.1
invoke~=2.2.0